generate: check transitive availability of optional dependencies
This fixes e.g.,

depends(party, ?fun), depends(fun, alcohol), disable(alcohol)

Previously, as fun is available, it was added to party depends, but then
failed resolving the disabled alcohol.

Now, before adding fun as dep to party, it is recursively checked for
availability, and as alcohol is disabled, will be skipped.